refactor: event 발행 로직 클래스 분리
