+++ api-user/src/main/java/com/wellmeet/global/event/event/ReservationCreatedEvent.java
@@ -1,0 +1,31 @@
+package com.wellmeet.global.event.event;
+
+import com.wellmeet.domain.reservation.entity.Reservation;
+import java.time.LocalDateTime;
+import lombok.Getter;
+
+@Getter
+public class ReservationCreatedEvent {
+
+    private final Long reservationId;
+    private final String memberId;
+    private final String restaurantId;
+    private final String restaurantName;
+    private final String status;
+    private final int partySize;
+    private final String specialRequest;
+    private final LocalDateTime dateTime;
+    private final LocalDateTime createdAt;
+
+    public ReservationCreatedEvent(Reservation reservation) {
+        this.reservationId = reservation.getId();
+        this.memberId = reservation.getMember().getId();
+        this.restaurantId = reservation.getRestaurant().getId();
+        this.restaurantName = reservation.getRestaurantName();
+        this.status = reservation.getStatus().name();
+        this.partySize = reservation.getPartySize();
+        this.specialRequest = reservation.getSpecialRequest();
+        this.dateTime = reservation.getDateTime();
+        this.createdAt = reservation.getCreatedAt();
+    }
+}
